# docs/swagger.yaml
basePath: /api/v2/
definitions:
...
  model.EmailReq:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  model.LoginBody:
    properties:
      login_type:
        type: string
      uid:
        type: string
    required:
    - login_type
    - uid
    type: object
...
info:
  contact:
    email: admintool@investicstudio.com
    name: API Support
  description: API documentation for Crypto Studio Backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Crypto Studio Backend API
  version: "1.0"
paths:
  /auth/exist-email:
    post:
      consumes:
      - application/json
      description: Verifies if a given email is already registered
      parameters:
      - description: Email request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.EmailReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check if email exists
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your-token}" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"